[WIP] Add TLS to Docker-Swarm Template
This patch adds TLS at two key places. The first is on the Docker
daemon itself, which secures communication between the Swarm manager
service and the Docker Daemon. The second is on the Swarm manager,
which secures communication between the Magnum conductor and the
Swarm manager.

TODO:
 * Add docs for generating a client key/csr/crt
 * Split out bug fixes into separate commits
 * Find better way to handle json from make-certs.sh?
 * Use trust token instead of user's token

Partial-Implements: blueprint secure-docker
Change-Id: Ic88edf4b2e0005f6aa0a6df33b94ff275a5623d2