Fix swarm bay failure reporting
The old method of detecting failures was very likely to fail in
many cases because it relied on all bay services being started by
the time cloud-init finished. This is a problem because the bay
services are started asynchronously and can take quite a while to
start.

The new method relies on systemd's OnFailure directive to kick off
specific service units when a failure is detected. Both the swarm
agent and manager have their own failure service so that we are
not overloading a single wait condition with multiple potential
failures.

Closes-bug: #1502329